ai-service/requirements.txt
# E-KYC AI Service Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML & Computer Vision
opencv-python==4.9.0.80
mediapipe==0.10.9
easyocr==1.7.1
numpy==1.26.3
Pillow==10.2.0

# Face Verification (FaceNet)
facenet-pytorch==2.5.3
torch==2.1.2
torchvision==0.16.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
